+++ srs/rtc_test.go
@@ -22,19 +22,106 @@ package srs
 
 import (
 	"context"
+	"flag"
 	"fmt"
+	"github.com/pion/transport/vnet"
+	"io"
+	"io/ioutil"
+	"math/rand"
+	"os"
+	"path"
+	"strings"
+	"sync"
+	"testing"
+	"time"
+
 	"github.com/ossrs/go-oryx-lib/errors"
 	"github.com/ossrs/go-oryx-lib/logger"
 	"github.com/pion/interceptor"
 	"github.com/pion/rtcp"
 	"github.com/pion/rtp"
-	"github.com/pion/transport/vnet"
-	"math/rand"
-	"os"
-	"sync"
-	"testing"
-	"time"
 )
+
+func prepareTest() error {
+	var err error
+
+	srsHttps = flag.Bool("srs-https", false, "Whther connect to HTTPS-API")
+	srsServer = flag.String("srs-server", "127.0.0.1", "The RTC server to connect to")
+	srsStream = flag.String("srs-stream", "/rtc/regression", "The RTC stream to play")
+	srsLog = flag.Bool("srs-log", false, "Whether enable the detail log")
+	srsTimeout = flag.Int("srs-timeout", 5000, "For each case, the timeout in ms")
+	srsPlayPLI = flag.Int("srs-play-pli", 5000, "The PLI interval in seconds for player.")
+	srsPlayOKPackets = flag.Int("srs-play-ok-packets", 10, "If recv N RTP packets, it's ok, or fail")
+	srsPublishOKPackets = flag.Int("srs-publish-ok-packets", 3, "If send N RTP, recv N RTCP packets, it's ok, or fail")
+	srsPublishAudio = flag.String("srs-publish-audio", "avatar.ogg", "The audio file for publisher.")
+	srsPublishVideo = flag.String("srs-publish-video", "avatar.h264", "The video file for publisher.")
+	srsPublishVideoFps = flag.Int("srs-publish-video-fps", 25, "The video fps for publisher.")
+	srsVnetClientIP = flag.String("srs-vnet-client-ip", "192.168.168.168", "The client ip in pion/vnet.")
+	srsDTLSDropPackets = flag.Int("srs-dtls-drop-packets", 5, "If dropped N packets, it's ok, or fail")
+
+	// Should parse it first.
+	flag.Parse()
+
+	// The stream should starts with /, for example, /rtc/regression
+	if !strings.HasPrefix(*srsStream, "/") {
+		*srsStream = "/" + *srsStream
+	}
+
+	// Generate srs protocol from whether use HTTPS.
+	srsSchema = "http"
+	if *srsHttps {
+		srsSchema = "https"
+	}
+
+	// Check file.
+	tryOpenFile := func(filename string) (string, error) {
+		if filename == "" {
+			return filename, nil
+		}
+
+		f, err := os.Open(filename)
+		if err != nil {
+			nfilename := path.Join("../", filename)
+			f2, err := os.Open(nfilename)
+			if err != nil {
+				return filename, errors.Wrapf(err, "No video file at %v or %v", filename, nfilename)
+			}
+			defer f2.Close()
+
+			return nfilename, nil
+		}
+		defer f.Close()
+
+		return filename, nil
+	}
+
+	if *srsPublishVideo, err = tryOpenFile(*srsPublishVideo); err != nil {
+		return err
+	}
+
+	if *srsPublishAudio, err = tryOpenFile(*srsPublishAudio); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func TestMain(m *testing.M) {
+	if err := prepareTest(); err != nil {
+		logger.Ef(nil, "Prepare test fail, err %+v", err)
+		os.Exit(-1)
+	}
+
+	// Disable the logger during all tests.
+	if *srsLog == false {
+		olw := logger.Switch(ioutil.Discard)
+		defer func() {
+			logger.Switch(olw)
+		}()
+	}
+
+	os.Exit(m.Run())
+}
 
 // Basic use scenario, publish a stream, then play it.
 func TestRtcBasic_PublishPlay(t *testing.T) {
@@ -50,12 +137,20 @@ func TestRtcBasic_PublishPlay(t *testing.T) {
 		}
 	}(ctx)
 
+	var resources []io.Closer
+	defer func() {
+		for _, resource := range resources {
+			resource.Close()
+		}
+	}()
+
 	var wg sync.WaitGroup
 	defer wg.Wait()
 
 	// The event notify.
 	var thePublisher *TestPublisher
 	var thePlayer *TestPlayer
+
 	mainReady, mainReadyCancel := context.WithCancel(context.Background())
 	publishReady, publishReadyCancel := context.WithCancel(context.Background())
 
@@ -66,76 +161,110 @@ func TestRtcBasic_PublishPlay(t *testing.T) {
 		defer cancel()
 
 		doInit := func() error {
-			playOK := *srsPlayOKPackets
-			vnetClientIP := *srsVnetClientIP
-
-			// Create top level test object.
-			api, err := NewTestWebRTCAPI()
-			if err != nil {
-				return err
-			}
-			defer api.Close()
-
+			playOK, vnetClientIP := *srsPlayOKPackets, *srsVnetClientIP
 			streamSuffix := fmt.Sprintf("basic-publish-play-%v-%v", os.Getpid(), rand.Int())
-			play := NewTestPlayer(api, func(play *TestPlayer) {
+
+			// Initialize player with private api.
+			if play, err := NewTestPlayer(nil, func(play *TestPlayer) error {
 				play.streamSuffix = streamSuffix
-			})
-			defer play.Close()
+				resources = append(resources, play)
 
-			pub := NewTestPublisher(api, func(pub *TestPublisher) {
-				pub.streamSuffix = streamSuffix
-				pub.iceReadyCancel = publishReadyCancel
-			})
-			defer pub.Close()
+				api, err := NewTestWebRTCAPI()
+				if err != nil {
+					return err
+				}
+				resources = append(resources, api)
+				play.api = api
 
-			if err := api.Setup(vnetClientIP, func(api *TestWebRTCAPI) {
-				var nnWriteRTP, nnReadRTP, nnWriteRTCP, nnReadRTCP int64
-				api.registry.Add(NewRTPInterceptor(func(i *RTPInterceptor) {
-					i.rtpReader = func(buf []byte, attributes interceptor.Attributes) (int, interceptor.Attributes, error) {
-						nn, attr, err := i.nextRTPReader.Read(buf, attributes)
-						nnReadRTP++
-						return nn, attr, err
-					}
-					i.rtpWriter = func(header *rtp.Header, payload []byte, attributes interceptor.Attributes) (int, error) {
-						nn, err := i.nextRTPWriter.Write(header, payload, attributes)
-
-						nnWriteRTP++
-						logger.Tf(ctx, "publish rtp=(read:%v write:%v), rtcp=(read:%v write:%v) packets",
-							nnReadRTP, nnWriteRTP, nnReadRTCP, nnWriteRTCP)
-						return nn, err
-					}
-				}))
-				api.registry.Add(NewRTCPInterceptor(func(i *RTCPInterceptor) {
-					i.rtcpReader = func(buf []byte, attributes interceptor.Attributes) (int, interceptor.Attributes, error) {
-						nn, attr, err := i.nextRTCPReader.Read(buf, attributes)
-						nnReadRTCP++
-						return nn, attr, err
-					}
-					i.rtcpWriter = func(pkts []rtcp.Packet, attributes interceptor.Attributes) (int, error) {
-						nn, err := i.nextRTCPWriter.Write(pkts, attributes)
-						nnWriteRTCP++
-						return nn, err
-					}
-				}))
-			}, func(api *TestWebRTCAPI) {
-				var nn uint64
-				api.registry.Add(NewRTPInterceptor(func(i *RTPInterceptor) {
-					i.rtpReader = func(payload []byte, attributes interceptor.Attributes) (int, interceptor.Attributes, error) {
-						if nn++; nn >= uint64(playOK) {
-							cancel() // Completed.
+				var nnPlayWriteRTCP, nnPlayReadRTCP, nnPlayWriteRTP, nnPlayReadRTP uint64
+				if err := api.Setup(vnetClientIP, func(api *TestWebRTCAPI) {
+					api.registry.Add(NewRTPInterceptor(func(i *RTPInterceptor) {
+						i.rtpReader = func(payload []byte, attributes interceptor.Attributes) (int, interceptor.Attributes, error) {
+							if nnPlayReadRTP++; nnPlayReadRTP >= uint64(playOK) {
+								cancel() // Completed.
+							}
+							logger.Tf(ctx, "Play rtp=(recv:%v, send:%v), rtcp=(recv:%v send:%v) packets",
+								nnPlayReadRTP, nnPlayWriteRTP, nnPlayReadRTCP, nnPlayWriteRTCP)
+							return i.nextRTPReader.Read(payload, attributes)
 						}
-						logger.Tf(ctx, "play got %v packets", nn)
-						return i.nextRTPReader.Read(payload, attributes)
-					}
-				}))
+					}))
+					api.registry.Add(NewRTCPInterceptor(func(i *RTCPInterceptor) {
+						i.rtcpReader = func(buf []byte, attributes interceptor.Attributes) (int, interceptor.Attributes, error) {
+							nn, attr, err := i.nextRTCPReader.Read(buf, attributes)
+							nnPlayReadRTCP++
+							return nn, attr, err
+						}
+						i.rtcpWriter = func(pkts []rtcp.Packet, attributes interceptor.Attributes) (int, error) {
+							nn, err := i.nextRTCPWriter.Write(pkts, attributes)
+							nnPlayWriteRTCP++
+							return nn, err
+						}
+					}))
+				}); err != nil {
+					return err
+				}
+
+				return nil
 			}); err != nil {
 				return err
+			} else {
+				thePlayer = play
 			}
 
-			// Set the available objects.
+			// Initialize publisher with private api.
+			if pub, err := NewTestPublisher(nil, func(pub *TestPublisher) error {
+				pub.streamSuffix = streamSuffix
+				pub.iceReadyCancel = publishReadyCancel
+				resources = append(resources, pub)
+
+				api, err := NewTestWebRTCAPI()
+				if err != nil {
+					return err
+				}
+				resources = append(resources, api)
+				pub.api = api
+
+				var nnPubWriteRTCP, nnPubReadRTCP, nnPubWriteRTP, nnPubReadRTP uint64
+				if err := api.Setup(vnetClientIP, func(api *TestWebRTCAPI) {
+					api.registry.Add(NewRTPInterceptor(func(i *RTPInterceptor) {
+						i.rtpReader = func(buf []byte, attributes interceptor.Attributes) (int, interceptor.Attributes, error) {
+							nn, attr, err := i.nextRTPReader.Read(buf, attributes)
+							nnPubReadRTP++
+							return nn, attr, err
+						}
+						i.rtpWriter = func(header *rtp.Header, payload []byte, attributes interceptor.Attributes) (int, error) {
+							nn, err := i.nextRTPWriter.Write(header, payload, attributes)
+							nnPubWriteRTP++
+							logger.Tf(ctx, "Publish rtp=(recv:%v, send:%v), rtcp=(recv:%v send:%v) packets",
+								nnPubReadRTP, nnPubWriteRTP, nnPubReadRTCP, nnPubWriteRTCP)
+							return nn, err
+						}
+					}))
+					api.registry.Add(NewRTCPInterceptor(func(i *RTCPInterceptor) {
+						i.rtcpReader = func(buf []byte, attributes interceptor.Attributes) (int, interceptor.Attributes, error) {
+							nn, attr, err := i.nextRTCPReader.Read(buf, attributes)
+							nnPubReadRTCP++
+							return nn, attr, err
+						}
+						i.rtcpWriter = func(pkts []rtcp.Packet, attributes interceptor.Attributes) (int, error) {
+							nn, err := i.nextRTCPWriter.Write(pkts, attributes)
+							nnPubWriteRTCP++
+							return nn, err
+						}
+					}))
+				}); err != nil {
+					return err
+				}
+
+				return nil
+			}); err != nil {
+				return err
+			} else {
+				thePublisher = pub
+			}
+
+			// Init done.
 			mainReadyCancel()
-			thePublisher = pub
-			thePlayer = play
 
 			<-ctx.Done()
 			return nil
@@ -158,17 +287,10 @@ func TestRtcBasic_PublishPlay(t *testing.T) {
 		case <-mainReady.Done():
 		}
 
-		doPublish := func() error {
-			if err := thePublisher.Run(logger.WithContext(ctx), cancel); err != nil {
-				return err
-			}
-
-			logger.Tf(ctx, "pub done")
-			return nil
-		}
-		if err := doPublish(); err != nil {
+		if err := thePublisher.Run(logger.WithContext(ctx), cancel); err != nil {
 			r2 = err
 		}
+		logger.Tf(ctx, "pub done")
 	}()
 
 	// Run player.
@@ -180,27 +302,13 @@ func TestRtcBasic_PublishPlay(t *testing.T) {
 		select {
 		case <-ctx.Done():
 			return
-		case <-mainReady.Done():
-		}
-
-		select {
-		case <-ctx.Done():
-			return
 		case <-publishReady.Done():
 		}
 
-		doPlay := func() error {
-			if err := thePlayer.Run(logger.WithContext(ctx), cancel); err != nil {
-				return err
-			}
-
-			logger.Tf(ctx, "play done")
-			return nil
-		}
-		if err := doPlay(); err != nil {
+		if err := thePlayer.Run(logger.WithContext(ctx), cancel); err != nil {
 			r3 = err
 		}
-
+		logger.Tf(ctx, "play done")
 	}()
 }
 
@@ -222,21 +330,31 @@ func TestRtcDTLS_ClientActive_Default(t *testing.T) {
 		defer api.Close()
 
 		streamSuffix := fmt.Sprintf("dtls-passive-no-arq-%v-%v", os.Getpid(), rand.Int())
-		p := NewTestPublisher(api, func(p *TestPublisher) {
+		p, err := NewTestPublisher(api, func(p *TestPublisher) error {
 			p.streamSuffix = streamSuffix
 			p.onOffer = testUtilSetupActive
+			return nil
 		})
+		if err != nil {
+			return err
+		}
 		defer p.Close()
 
 		if err := api.Setup(vnetClientIP, func(api *TestWebRTCAPI) {
-			var nn int64
+			var nnRTCP, nnRTP int64
 			api.registry.Add(NewRTPInterceptor(func(i *RTPInterceptor) {
 				i.rtpWriter = func(header *rtp.Header, payload []byte, attributes interceptor.Attributes) (int, error) {
-					if nn++; nn >= int64(publishOK) {
+					nnRTP++
+					return i.nextRTPWriter.Write(header, payload, attributes)
+				}
+			}))
+			api.registry.Add(NewRTCPInterceptor(func(i *RTCPInterceptor) {
+				i.rtcpReader = func(buf []byte, attributes interceptor.Attributes) (int, interceptor.Attributes, error) {
+					if nnRTCP++; nnRTCP >= int64(publishOK) && nnRTP >= int64(publishOK) {
 						cancel() // Send enough packets, done.
 					}
-					logger.Tf(ctx, "publish write %v packets", nn)
-					return i.nextRTPWriter.Write(header, payload, attributes)
+					logger.Tf(ctx, "publish write %v RTP read %v RTCP packets", nnRTP, nnRTCP)
+					return i.nextRTCPReader.Read(buf, attributes)
 				}
 			}))
 		}, func(api *TestWebRTCAPI) {
@@ -276,21 +394,31 @@ func TestRtcDTLS_ClientPassive_Default(t *testing.T) {
 		defer api.Close()
 
 		streamSuffix := fmt.Sprintf("dtls-active-no-arq-%v-%v", os.Getpid(), rand.Int())
-		p := NewTestPublisher(api, func(p *TestPublisher) {
+		p, err := NewTestPublisher(api, func(p *TestPublisher) error {
 			p.streamSuffix = streamSuffix
 			p.onOffer = testUtilSetupPassive
+			return nil
 		})
+		if err != nil {
+			return err
+		}
 		defer p.Close()
 
 		if err := api.Setup(vnetClientIP, func(api *TestWebRTCAPI) {
-			var nn int64
+			var nnRTCP, nnRTP int64
 			api.registry.Add(NewRTPInterceptor(func(i *RTPInterceptor) {
 				i.rtpWriter = func(header *rtp.Header, payload []byte, attributes interceptor.Attributes) (int, error) {
-					if nn++; nn >= int64(publishOK) {
+					nnRTP++
+					return i.nextRTPWriter.Write(header, payload, attributes)
+				}
+			}))
+			api.registry.Add(NewRTCPInterceptor(func(i *RTCPInterceptor) {
+				i.rtcpReader = func(buf []byte, attributes interceptor.Attributes) (int, interceptor.Attributes, error) {
+					if nnRTCP++; nnRTCP >= int64(publishOK) && nnRTP >= int64(publishOK) {
 						cancel() // Send enough packets, done.
 					}
-					logger.Tf(ctx, "publish write %v packets", nn)
-					return i.nextRTPWriter.Write(header, payload, attributes)
+					logger.Tf(ctx, "publish write %v RTP read %v RTCP packets", nnRTP, nnRTCP)
+					return i.nextRTCPReader.Read(buf, attributes)
 				}
 			}))
 		}, func(api *TestWebRTCAPI) {
@@ -327,21 +455,31 @@ func TestRtcDTLS_ClientActive_Duplicated_Alert(t *testing.T) {
 		defer api.Close()
 
 		streamSuffix := fmt.Sprintf("dtls-active-no-arq-%v-%v", os.Getpid(), rand.Int())
-		p := NewTestPublisher(api, func(p *TestPublisher) {
+		p, err := NewTestPublisher(api, func(p *TestPublisher) error {
 			p.streamSuffix = streamSuffix
 			p.onOffer = testUtilSetupActive
+			return nil
 		})
+		if err != nil {
+			return err
+		}
 		defer p.Close()
 
 		if err := api.Setup(vnetClientIP, func(api *TestWebRTCAPI) {
-			var nn int64
+			var nnRTCP, nnRTP int64
 			api.registry.Add(NewRTPInterceptor(func(i *RTPInterceptor) {
 				i.rtpWriter = func(header *rtp.Header, payload []byte, attributes interceptor.Attributes) (int, error) {
-					if nn++; nn >= int64(publishOK) {
+					nnRTP++
+					return i.nextRTPWriter.Write(header, payload, attributes)
+				}
+			}))
+			api.registry.Add(NewRTCPInterceptor(func(i *RTCPInterceptor) {
+				i.rtcpReader = func(buf []byte, attributes interceptor.Attributes) (int, interceptor.Attributes, error) {
+					if nnRTCP++; nnRTCP >= int64(publishOK) && nnRTP >= int64(publishOK) {
 						cancel() // Send enough packets, done.
 					}
-					logger.Tf(ctx, "publish write %v packets", nn)
-					return i.nextRTPWriter.Write(header, payload, attributes)
+					logger.Tf(ctx, "publish write %v RTP read %v RTCP packets", nnRTP, nnRTCP)
+					return i.nextRTCPReader.Read(buf, attributes)
 				}
 			}))
 		}, func(api *TestWebRTCAPI) {
@@ -385,21 +523,31 @@ func TestRtcDTLS_ClientPassive_Duplicated_Alert(t *testing.T) {
 		defer api.Close()
 
 		streamSuffix := fmt.Sprintf("dtls-active-no-arq-%v-%v", os.Getpid(), rand.Int())
-		p := NewTestPublisher(api, func(p *TestPublisher) {
+		p, err := NewTestPublisher(api, func(p *TestPublisher) error {
 			p.streamSuffix = streamSuffix
 			p.onOffer = testUtilSetupPassive
+			return nil
 		})
+		if err != nil {
+			return err
+		}
 		defer p.Close()
 
 		if err := api.Setup(vnetClientIP, func(api *TestWebRTCAPI) {
-			var nn int64
+			var nnRTCP, nnRTP int64
 			api.registry.Add(NewRTPInterceptor(func(i *RTPInterceptor) {
 				i.rtpWriter = func(header *rtp.Header, payload []byte, attributes interceptor.Attributes) (int, error) {
-					if nn++; nn >= int64(publishOK) {
+					nnRTP++
+					return i.nextRTPWriter.Write(header, payload, attributes)
+				}
+			}))
+			api.registry.Add(NewRTCPInterceptor(func(i *RTCPInterceptor) {
+				i.rtcpReader = func(buf []byte, attributes interceptor.Attributes) (int, interceptor.Attributes, error) {
+					if nnRTCP++; nnRTCP >= int64(publishOK) && nnRTP >= int64(publishOK) {
 						cancel() // Send enough packets, done.
 					}
-					logger.Tf(ctx, "publish write %v packets", nn)
-					return i.nextRTPWriter.Write(header, payload, attributes)
+					logger.Tf(ctx, "publish write %v RTP read %v RTCP packets", nnRTP, nnRTCP)
+					return i.nextRTCPReader.Read(buf, attributes)
 				}
 			}))
 		}, func(api *TestWebRTCAPI) {
@@ -450,21 +598,31 @@ func TestRtcDTLS_ClientActive_ARQ_ClientHello_ByDropped_ClientHello(t *testing.T
 		defer api.Close()
 
 		streamSuffix := fmt.Sprintf("dtls-active-arq-client-hello-%v-%v", os.Getpid(), rand.Int())
-		p := NewTestPublisher(api, func(p *TestPublisher) {
+		p, err := NewTestPublisher(api, func(p *TestPublisher) error {
 			p.streamSuffix = streamSuffix
 			p.onOffer = testUtilSetupActive
+			return nil
 		})
+		if err != nil {
+			return err
+		}
 		defer p.Close()
 
 		if err := api.Setup(vnetClientIP, func(api *TestWebRTCAPI) {
-			var nn int64
+			var nnRTCP, nnRTP int64
 			api.registry.Add(NewRTPInterceptor(func(i *RTPInterceptor) {
 				i.rtpWriter = func(header *rtp.Header, payload []byte, attributes interceptor.Attributes) (int, error) {
-					if nn++; nn >= int64(publishOK) {
+					nnRTP++
+					return i.nextRTPWriter.Write(header, payload, attributes)
+				}
+			}))
+			api.registry.Add(NewRTCPInterceptor(func(i *RTCPInterceptor) {
+				i.rtcpReader = func(buf []byte, attributes interceptor.Attributes) (int, interceptor.Attributes, error) {
+					if nnRTCP++; nnRTCP >= int64(publishOK) && nnRTP >= int64(publishOK) {
 						cancel() // Send enough packets, done.
 					}
-					logger.Tf(ctx, "publish write %v packets", nn)
-					return i.nextRTPWriter.Write(header, payload, attributes)
+					logger.Tf(ctx, "publish write %v RTP read %v RTCP packets", nnRTP, nnRTCP)
+					return i.nextRTCPReader.Read(buf, attributes)
 				}
 			}))
 		}, func(api *TestWebRTCAPI) {
@@ -527,21 +685,31 @@ func TestRtcDTLS_ClientPassive_ARQ_ClientHello_ByDropped_ClientHello(t *testing.
 		defer api.Close()
 
 		streamSuffix := fmt.Sprintf("dtls-passive-arq-client-hello-%v-%v", os.Getpid(), rand.Int())
-		p := NewTestPublisher(api, func(p *TestPublisher) {
+		p, err := NewTestPublisher(api, func(p *TestPublisher) error {
 			p.streamSuffix = streamSuffix
 			p.onOffer = testUtilSetupPassive
+			return nil
 		})
+		if err != nil {
+			return err
+		}
 		defer p.Close()
 
 		if err := api.Setup(vnetClientIP, func(api *TestWebRTCAPI) {
-			var nn int64
+			var nnRTCP, nnRTP int64
 			api.registry.Add(NewRTPInterceptor(func(i *RTPInterceptor) {
 				i.rtpWriter = func(header *rtp.Header, payload []byte, attributes interceptor.Attributes) (int, error) {
-					if nn++; nn >= int64(publishOK) {
+					nnRTP++
+					return i.nextRTPWriter.Write(header, payload, attributes)
+				}
+			}))
+			api.registry.Add(NewRTCPInterceptor(func(i *RTCPInterceptor) {
+				i.rtcpReader = func(buf []byte, attributes interceptor.Attributes) (int, interceptor.Attributes, error) {
+					if nnRTCP++; nnRTCP >= int64(publishOK) && nnRTP >= int64(publishOK) {
 						cancel() // Send enough packets, done.
 					}
-					logger.Tf(ctx, "publish write %v packets", nn)
-					return i.nextRTPWriter.Write(header, payload, attributes)
+					logger.Tf(ctx, "publish write %v RTP read %v RTCP packets", nnRTP, nnRTCP)
+					return i.nextRTCPReader.Read(buf, attributes)
 				}
 			}))
 		}, func(api *TestWebRTCAPI) {
@@ -603,21 +771,31 @@ func TestRtcDTLS_ClientActive_ARQ_ClientHello_ByDropped_ServerHello(t *testing.T
 		defer api.Close()
 
 		streamSuffix := fmt.Sprintf("dtls-active-arq-client-hello-%v-%v", os.Getpid(), rand.Int())
-		p := NewTestPublisher(api, func(p *TestPublisher) {
+		p, err := NewTestPublisher(api, func(p *TestPublisher) error {
 			p.streamSuffix = streamSuffix
 			p.onOffer = testUtilSetupActive
+			return nil
 		})
+		if err != nil {
+			return err
+		}
 		defer p.Close()
 
 		if err := api.Setup(vnetClientIP, func(api *TestWebRTCAPI) {
-			var nn int64
+			var nnRTCP, nnRTP int64
 			api.registry.Add(NewRTPInterceptor(func(i *RTPInterceptor) {
 				i.rtpWriter = func(header *rtp.Header, payload []byte, attributes interceptor.Attributes) (int, error) {
-					if nn++; nn >= int64(publishOK) {
+					nnRTP++
+					return i.nextRTPWriter.Write(header, payload, attributes)
+				}
+			}))
+			api.registry.Add(NewRTCPInterceptor(func(i *RTCPInterceptor) {
+				i.rtcpReader = func(buf []byte, attributes interceptor.Attributes) (int, interceptor.Attributes, error) {
+					if nnRTCP++; nnRTCP >= int64(publishOK) && nnRTP >= int64(publishOK) {
 						cancel() // Send enough packets, done.
 					}
-					logger.Tf(ctx, "publish write %v packets", nn)
-					return i.nextRTPWriter.Write(header, payload, attributes)
+					logger.Tf(ctx, "publish write %v RTP read %v RTCP packets", nnRTP, nnRTCP)
+					return i.nextRTCPReader.Read(buf, attributes)
 				}
 			}))
 		}, func(api *TestWebRTCAPI) {
@@ -690,21 +868,31 @@ func TestRtcDTLS_ClientPassive_ARQ_ClientHello_ByDropped_ServerHello(t *testing.
 		defer api.Close()
 
 		streamSuffix := fmt.Sprintf("dtls-passive-arq-client-hello-%v-%v", os.Getpid(), rand.Int())
-		p := NewTestPublisher(api, func(p *TestPublisher) {
+		p, err := NewTestPublisher(api, func(p *TestPublisher) error {
 			p.streamSuffix = streamSuffix
 			p.onOffer = testUtilSetupPassive
+			return nil
 		})
+		if err != nil {
+			return err
+		}
 		defer p.Close()
 
 		if err := api.Setup(vnetClientIP, func(api *TestWebRTCAPI) {
-			var nn int64
+			var nnRTCP, nnRTP int64
 			api.registry.Add(NewRTPInterceptor(func(i *RTPInterceptor) {
 				i.rtpWriter = func(header *rtp.Header, payload []byte, attributes interceptor.Attributes) (int, error) {
-					if nn++; nn >= int64(publishOK) {
+					nnRTP++
+					return i.nextRTPWriter.Write(header, payload, attributes)
+				}
+			}))
+			api.registry.Add(NewRTCPInterceptor(func(i *RTCPInterceptor) {
+				i.rtcpReader = func(buf []byte, attributes interceptor.Attributes) (int, interceptor.Attributes, error) {
+					if nnRTCP++; nnRTCP >= int64(publishOK) && nnRTP >= int64(publishOK) {
 						cancel() // Send enough packets, done.
 					}
-					logger.Tf(ctx, "publish write %v packets", nn)
-					return i.nextRTPWriter.Write(header, payload, attributes)
+					logger.Tf(ctx, "publish write %v RTP read %v RTCP packets", nnRTP, nnRTCP)
+					return i.nextRTCPReader.Read(buf, attributes)
 				}
 			}))
 		}, func(api *TestWebRTCAPI) {
@@ -774,21 +962,31 @@ func TestRtcDTLS_ClientActive_ARQ_Certificate_ByDropped_Certificate(t *testing.T
 		defer api.Close()
 
 		streamSuffix := fmt.Sprintf("dtls-active-arq-certificate-%v-%v", os.Getpid(), rand.Int())
-		p := NewTestPublisher(api, func(p *TestPublisher) {
+		p, err := NewTestPublisher(api, func(p *TestPublisher) error {
 			p.streamSuffix = streamSuffix
 			p.onOffer = testUtilSetupActive
+			return nil
 		})
+		if err != nil {
+			return err
+		}
 		defer p.Close()
 
 		if err := api.Setup(vnetClientIP, func(api *TestWebRTCAPI) {
-			var nn int64
+			var nnRTCP, nnRTP int64
 			api.registry.Add(NewRTPInterceptor(func(i *RTPInterceptor) {
 				i.rtpWriter = func(header *rtp.Header, payload []byte, attributes interceptor.Attributes) (int, error) {
-					if nn++; nn >= int64(publishOK) {
+					nnRTP++
+					return i.nextRTPWriter.Write(header, payload, attributes)
+				}
+			}))
+			api.registry.Add(NewRTCPInterceptor(func(i *RTCPInterceptor) {
+				i.rtcpReader = func(buf []byte, attributes interceptor.Attributes) (int, interceptor.Attributes, error) {
+					if nnRTCP++; nnRTCP >= int64(publishOK) && nnRTP >= int64(publishOK) {
 						cancel() // Send enough packets, done.
 					}
-					logger.Tf(ctx, "publish write %v packets", nn)
-					return i.nextRTPWriter.Write(header, payload, attributes)
+					logger.Tf(ctx, "publish write %v RTP read %v RTCP packets", nnRTP, nnRTCP)
+					return i.nextRTCPReader.Read(buf, attributes)
 				}
 			}))
 		}, func(api *TestWebRTCAPI) {
@@ -850,21 +1048,31 @@ func TestRtcDTLS_ClientPassive_ARQ_Certificate_ByDropped_Certificate(t *testing.
 		defer api.Close()
 
 		streamSuffix := fmt.Sprintf("dtls-passive-arq-certificate-%v-%v", os.Getpid(), rand.Int())
-		p := NewTestPublisher(api, func(p *TestPublisher) {
+		p, err := NewTestPublisher(api, func(p *TestPublisher) error {
 			p.streamSuffix = streamSuffix
 			p.onOffer = testUtilSetupPassive
+			return nil
 		})
+		if err != nil {
+			return err
+		}
 		defer p.Close()
 
 		if err := api.Setup(vnetClientIP, func(api *TestWebRTCAPI) {
-			var nn int64
+			var nnRTCP, nnRTP int64
 			api.registry.Add(NewRTPInterceptor(func(i *RTPInterceptor) {
 				i.rtpWriter = func(header *rtp.Header, payload []byte, attributes interceptor.Attributes) (int, error) {
-					if nn++; nn >= int64(publishOK) {
+					nnRTP++
+					return i.nextRTPWriter.Write(header, payload, attributes)
+				}
+			}))
+			api.registry.Add(NewRTCPInterceptor(func(i *RTCPInterceptor) {
+				i.rtcpReader = func(buf []byte, attributes interceptor.Attributes) (int, interceptor.Attributes, error) {
+					if nnRTCP++; nnRTCP >= int64(publishOK) && nnRTP >= int64(publishOK) {
 						cancel() // Send enough packets, done.
 					}
-					logger.Tf(ctx, "publish write %v packets", nn)
-					return i.nextRTPWriter.Write(header, payload, attributes)
+					logger.Tf(ctx, "publish write %v RTP read %v RTCP packets", nnRTP, nnRTCP)
+					return i.nextRTCPReader.Read(buf, attributes)
 				}
 			}))
 		}, func(api *TestWebRTCAPI) {
@@ -926,21 +1134,31 @@ func TestRtcDTLS_ClientActive_ARQ_Certificate_ByDropped_ChangeCipherSpec(t *test
 		defer api.Close()
 
 		streamSuffix := fmt.Sprintf("dtls-active-arq-certificate-%v-%v", os.Getpid(), rand.Int())
-		p := NewTestPublisher(api, func(p *TestPublisher) {
+		p, err := NewTestPublisher(api, func(p *TestPublisher) error {
 			p.streamSuffix = streamSuffix
 			p.onOffer = testUtilSetupActive
+			return nil
 		})
+		if err != nil {
+			return err
+		}
 		defer p.Close()
 
 		if err := api.Setup(vnetClientIP, func(api *TestWebRTCAPI) {
-			var nn int64
+			var nnRTCP, nnRTP int64
 			api.registry.Add(NewRTPInterceptor(func(i *RTPInterceptor) {
 				i.rtpWriter = func(header *rtp.Header, payload []byte, attributes interceptor.Attributes) (int, error) {
-					if nn++; nn >= int64(publishOK) {
+					nnRTP++
+					return i.nextRTPWriter.Write(header, payload, attributes)
+				}
+			}))
+			api.registry.Add(NewRTCPInterceptor(func(i *RTCPInterceptor) {
+				i.rtcpReader = func(buf []byte, attributes interceptor.Attributes) (int, interceptor.Attributes, error) {
+					if nnRTCP++; nnRTCP >= int64(publishOK) && nnRTP >= int64(publishOK) {
 						cancel() // Send enough packets, done.
 					}
-					logger.Tf(ctx, "publish write %v packets", nn)
-					return i.nextRTPWriter.Write(header, payload, attributes)
+					logger.Tf(ctx, "publish write %v RTP read %v RTCP packets", nnRTP, nnRTCP)
+					return i.nextRTCPReader.Read(buf, attributes)
 				}
 			}))
 		}, func(api *TestWebRTCAPI) {
@@ -1011,21 +1229,31 @@ func TestRtcDTLS_ClientPassive_ARQ_Certificate_ByDropped_ChangeCipherSpec(t *tes
 		defer api.Close()
 
 		streamSuffix := fmt.Sprintf("dtls-passive-arq-certificate-%v-%v", os.Getpid(), rand.Int())
-		p := NewTestPublisher(api, func(p *TestPublisher) {
+		p, err := NewTestPublisher(api, func(p *TestPublisher) error {
 			p.streamSuffix = streamSuffix
 			p.onOffer = testUtilSetupPassive
+			return nil
 		})
+		if err != nil {
+			return err
+		}
 		defer p.Close()
 
 		if err := api.Setup(vnetClientIP, func(api *TestWebRTCAPI) {
-			var nn int64
+			var nnRTCP, nnRTP int64
 			api.registry.Add(NewRTPInterceptor(func(i *RTPInterceptor) {
 				i.rtpWriter = func(header *rtp.Header, payload []byte, attributes interceptor.Attributes) (int, error) {
-					if nn++; nn >= int64(publishOK) {
+					nnRTP++
+					return i.nextRTPWriter.Write(header, payload, attributes)
+				}
+			}))
+			api.registry.Add(NewRTCPInterceptor(func(i *RTCPInterceptor) {
+				i.rtcpReader = func(buf []byte, attributes interceptor.Attributes) (int, interceptor.Attributes, error) {
+					if nnRTCP++; nnRTCP >= int64(publishOK) && nnRTP >= int64(publishOK) {
 						cancel() // Send enough packets, done.
 					}
-					logger.Tf(ctx, "publish write %v packets", nn)
-					return i.nextRTPWriter.Write(header, payload, attributes)
+					logger.Tf(ctx, "publish write %v RTP read %v RTCP packets", nnRTP, nnRTCP)
+					return i.nextRTCPReader.Read(buf, attributes)
 				}
 			}))
 		}, func(api *TestWebRTCAPI) {
@@ -1087,10 +1315,14 @@ func TestRtcDTLS_ClientPassive_ARQ_DropAllAfter_ClientHello(t *testing.T) {
 		defer api.Close()
 
 		streamSuffix := fmt.Sprintf("dtls-passive-no-arq-%v-%v", os.Getpid(), rand.Int())
-		p := NewTestPublisher(api, func(p *TestPublisher) {
+		p, err := NewTestPublisher(api, func(p *TestPublisher) error {
 			p.streamSuffix = streamSuffix
 			p.onOffer = testUtilSetupPassive
+			return nil
 		})
+		if err != nil {
+			return err
+		}
 		defer p.Close()
 
 		if err := api.Setup(vnetClientIP, func(api *TestWebRTCAPI) {
@@ -1145,10 +1377,14 @@ func TestRtcDTLS_ClientPassive_ARQ_DropAllAfter_ServerHello(t *testing.T) {
 		defer api.Close()
 
 		streamSuffix := fmt.Sprintf("dtls-passive-no-arq-%v-%v", os.Getpid(), rand.Int())
-		p := NewTestPublisher(api, func(p *TestPublisher) {
+		p, err := NewTestPublisher(api, func(p *TestPublisher) error {
 			p.streamSuffix = streamSuffix
 			p.onOffer = testUtilSetupPassive
+			return nil
 		})
+		if err != nil {
+			return err
+		}
 		defer p.Close()
 
 		if err := api.Setup(vnetClientIP, func(api *TestWebRTCAPI) {
@@ -1203,10 +1439,14 @@ func TestRtcDTLS_ClientPassive_ARQ_DropAllAfter_Certificate(t *testing.T) {
 		defer api.Close()
 
 		streamSuffix := fmt.Sprintf("dtls-passive-no-arq-%v-%v", os.Getpid(), rand.Int())
-		p := NewTestPublisher(api, func(p *TestPublisher) {
+		p, err := NewTestPublisher(api, func(p *TestPublisher) error {
 			p.streamSuffix = streamSuffix
 			p.onOffer = testUtilSetupPassive
+			return nil
 		})
+		if err != nil {
+			return err
+		}
 		defer p.Close()
 
 		if err := api.Setup(vnetClientIP, func(api *TestWebRTCAPI) {
@@ -1261,10 +1501,14 @@ func TestRtcDTLS_ClientPassive_ARQ_DropAllAfter_ChangeCipherSpec(t *testing.T) {
 		defer api.Close()
 
 		streamSuffix := fmt.Sprintf("dtls-passive-no-arq-%v-%v", os.Getpid(), rand.Int())
-		p := NewTestPublisher(api, func(p *TestPublisher) {
+		p, err := NewTestPublisher(api, func(p *TestPublisher) error {
 			p.streamSuffix = streamSuffix
 			p.onOffer = testUtilSetupPassive
+			return nil
 		})
+		if err != nil {
+			return err
+		}
 		defer p.Close()
 
 		if err := api.Setup(vnetClientIP, func(api *TestWebRTCAPI) {
@@ -1320,21 +1564,31 @@ func TestRtcDTLS_ClientPassive_ARQ_VeryBadNetwork(t *testing.T) {
 		defer api.Close()
 
 		streamSuffix := fmt.Sprintf("dtls-passive-no-arq-%v-%v", os.Getpid(), rand.Int())
-		p := NewTestPublisher(api, func(p *TestPublisher) {
+		p, err := NewTestPublisher(api, func(p *TestPublisher) error {
 			p.streamSuffix = streamSuffix
 			p.onOffer = testUtilSetupPassive
+			return nil
 		})
+		if err != nil {
+			return err
+		}
 		defer p.Close()
 
 		if err := api.Setup(vnetClientIP, func(api *TestWebRTCAPI) {
-			var nn int64
+			var nnRTCP, nnRTP int64
 			api.registry.Add(NewRTPInterceptor(func(i *RTPInterceptor) {
 				i.rtpWriter = func(header *rtp.Header, payload []byte, attributes interceptor.Attributes) (int, error) {
-					if nn++; nn >= int64(publishOK) {
+					nnRTP++
+					return i.nextRTPWriter.Write(header, payload, attributes)
+				}
+			}))
+			api.registry.Add(NewRTCPInterceptor(func(i *RTCPInterceptor) {
+				i.rtcpReader = func(buf []byte, attributes interceptor.Attributes) (int, interceptor.Attributes, error) {
+					if nnRTCP++; nnRTCP >= int64(publishOK) && nnRTP >= int64(publishOK) {
 						cancel() // Send enough packets, done.
 					}
-					logger.Tf(ctx, "publish write %v packets", nn)
-					return i.nextRTPWriter.Write(header, payload, attributes)
+					logger.Tf(ctx, "publish write %v RTP read %v RTCP packets", nnRTP, nnRTCP)
+					return i.nextRTCPReader.Read(buf, attributes)
 				}
 			}))
 		}, func(api *TestWebRTCAPI) {
@@ -1397,21 +1651,31 @@ func TestRtcDTLS_ClientPassive_ARQ_Certificate_After_ClientHello(t *testing.T) {
 		defer api.Close()
 
 		streamSuffix := fmt.Sprintf("dtls-passive-no-arq-%v-%v", os.Getpid(), rand.Int())
-		p := NewTestPublisher(api, func(p *TestPublisher) {
+		p, err := NewTestPublisher(api, func(p *TestPublisher) error {
 			p.streamSuffix = streamSuffix
 			p.onOffer = testUtilSetupPassive
+			return nil
 		})
+		if err != nil {
+			return err
+		}
 		defer p.Close()
 
 		if err := api.Setup(vnetClientIP, func(api *TestWebRTCAPI) {
-			var nn int64
+			var nnRTCP, nnRTP int64
 			api.registry.Add(NewRTPInterceptor(func(i *RTPInterceptor) {
 				i.rtpWriter = func(header *rtp.Header, payload []byte, attributes interceptor.Attributes) (int, error) {
-					if nn++; nn >= int64(publishOK) {
+					nnRTP++
+					return i.nextRTPWriter.Write(header, payload, attributes)
+				}
+			}))
+			api.registry.Add(NewRTCPInterceptor(func(i *RTCPInterceptor) {
+				i.rtcpReader = func(buf []byte, attributes interceptor.Attributes) (int, interceptor.Attributes, error) {
+					if nnRTCP++; nnRTCP >= int64(publishOK) && nnRTP >= int64(publishOK) {
 						cancel() // Send enough packets, done.
 					}
-					logger.Tf(ctx, "publish write %v packets", nn)
-					return i.nextRTPWriter.Write(header, payload, attributes)
+					logger.Tf(ctx, "publish write %v RTP read %v RTCP packets", nnRTP, nnRTCP)
+					return i.nextRTCPReader.Read(buf, attributes)
 				}
 			}))
 		}, func(api *TestWebRTCAPI) {
